Refactor: Extract version comparison logic to VersionUtils
Extracted the duplicated version comparison logic from `AppUpdateChecker` and `ProviderUpdateChecker` into a new `VersionUtils` object to improve code reuse and modularity. Also optimized performance by hoisting the regular expression compilation to a file-level private constant, avoiding redundant recompilations during loop iterations.
